chore(pages/MainPage): Abstract HomePage and BlogPage into own files
