+++ src/pages/BlogPage/BlogPage.tsx
@@ -1,0 +1,3 @@
+export const BlogPage = () => {
+  return <div> I am going to be a blog!!</div>;
+};
