+++ src/pages/MainPage/MainPage.tsx
@@ -2,30 +2,13 @@ import { useState } from "react";
 import type React from "react";
 
 import Box from "@mui/material/Box";
-import Grid2 from "@mui/material/Grid2";
 import Tab from "@mui/material/Tab";
 import Tabs from "@mui/material/Tabs";
 
-import { SideBySideSection, ThemeToggle } from "../../components";
-import { About, Contacts, Projects, SendEmail, Skills } from "../../sections";
-import type { Sections } from "../../types/sectionTypes";
+import { ThemeToggle } from "../../components";
+import { BlogPage, HomePage } from "../internal";
 
 import "./mainPageStyles.css";
-
-const sections: Sections = {
-  top: {
-    left: { component: About, size: 6, key: "about-section" },
-    right: { component: Skills, size: 4, key: "skills-section" },
-  },
-  middle: {
-    component: Projects,
-    key: "projects-section",
-  },
-  bottom: {
-    left: { component: SendEmail, size: 8, key: "certifications-section" },
-    right: { component: Contacts, size: 4, key: "contacts-section" },
-  },
-};
 
 interface TabPanelProps {
   children?: React.ReactNode;
@@ -33,15 +16,15 @@ interface TabPanelProps {
   value: number;
 }
 
-function CustomTabPanel(props: TabPanelProps) {
+function TabPanel(props: TabPanelProps) {
   const { children, value, index, ...other } = props;
 
   return (
     <div
       role="tabpanel"
       hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
+      id={`tabpanel-${index}`}
+      aria-labelledby={`tab-${index}`}
       {...other}
     >
       {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
@@ -51,8 +34,8 @@ function CustomTabPanel(props: TabPanelProps) {
 
 function a11yProps(index: number) {
   return {
-    id: `simple-tab-${index}`,
-    "aria-controls": `simple-tabpanel-${index}`,
+    id: `tab-${index}`,
+    "aria-controls": `tabpanel-${index}`,
   };
 }
 
@@ -63,8 +46,6 @@ export const MainPage: React.FC = () => {
     setValue(newValue);
   };
 
-  const ProjectsSection = sections.middle.component;
-
   return (
     <div className="full-width flex-grow-1">
       {/* Navigation bar -- start */}
@@ -73,10 +54,19 @@ export const MainPage: React.FC = () => {
           <Tabs
             value={value}
             onChange={handleChange}
-            aria-label="basic tabs example"
+            aria-label="Home and Blog tabs"
+            textColor="primary"
+            indicatorColor="primary"
           >
-            <Tab label="Home" {...a11yProps(0)} />
-            <Tab label="Blog" {...a11yProps(1)} />
+            <Tab
+              label={
+                <div>
+                  <span className="blue-font">~/</span>Home
+                </div>
+              }
+              {...a11yProps(0)}
+            />
+            <Tab label="/Blog" {...a11yProps(1)} />
           </Tabs>
         </Box>
         {/* ToggleTheme section */}
@@ -86,25 +76,13 @@ export const MainPage: React.FC = () => {
       </div>
       {/* Navigation bar -- end */}
 
-      {/* CustomTabPanels -- start */}
-      <CustomTabPanel value={value} index={0}>
-        {/* About and Skills section */}
-        <Grid2 className="styled-grid">
-          <SideBySideSection section={sections.top} />
-        </Grid2>
-        {/* Projects section */}
-        <Grid2 className="styled-grid">
-          <ProjectsSection />
-        </Grid2>
-        {/* Contacts section */}
-        <Grid2 className="styled-grid">
-          <SideBySideSection section={sections.bottom} />
-        </Grid2>
-      </CustomTabPanel>
-      <CustomTabPanel value={value} index={1}>
-        I am going to be a blog!!
-      </CustomTabPanel>
-      {/* CustomTabPanels -- end */}
+      {/* TabPanels */}
+      <TabPanel value={value} index={0}>
+        <HomePage />
+      </TabPanel>
+      <TabPanel value={value} index={1}>
+        <BlogPage />
+      </TabPanel>
     </div>
   );
 };
