+++ src/pages/internal.ts
@@ -1,0 +1,2 @@
+export { HomePage } from "./HomePage/HomePage";
+export { BlogPage } from "./BlogPage/BlogPage";
